+++ rs/lib/index.js
@@ -1,0 +1,3 @@
+var addon = require('../native');
+
+console.log(addon.generate(0, 100));
