+++ rs/native/Cargo.toml
@@ -18,3 +18,5 @@ neon = "0.3.3"
 strum = "0.11"
 strum_macros = "0.11"
 rand = "0.5"
+num-traits = "0.2"
+num-derive = "0.3"
