+++ rs/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "s",
+  "version": "0.1.0",
+  "description": "",
+  "main": "lib/index.js",
+  "author": "Eduardo Asafe <e@asafe.dev>",
+  "license": "MIT",
+  "dependencies": {
+    "neon-cli": "^0.3.3"
+  },
+  "scripts": {
+    "install": "neon build"
+  }
+}
